Streamlit-News-Article-Analysis/requirements.txt
ago==0.0.95
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==4.0.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.35
botocore==1.31.35
branca==0.6.0
bs4==0.0.1
cachetools==5.3.1
cchardet==2.1.7
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chromedriver-binary==118.0.5982.0.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
deprecation==2.1.0
dotmap==1.3.30
elastic-transport==8.4.0
elasticsearch==8.9.0
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.2.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
flatbuffers==23.5.26
folium==0.14.0
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.9.0
geographiclib==2.0
geopy==2.4.0
gitdb==4.0.10
GitPython==3.1.32
gotrue==1.0.4
greenlet==2.0.2
h11==0.14.0
hjson==3.1.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.25.1
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.292
langdetect==1.0.9
langsmith==0.0.27
libclang==16.0.6
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
news-fetch==0.2.8
news-please==1.5.33
newspaper3k==0.2.8
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.9
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
plac==1.3.5
platformdirs==3.10.0
postgrest==0.10.8
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.24.2
psutil==5.9.5
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
PyDispatcher==2.0.7
Pygments==2.16.1
Pympler==1.0.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
readability-lxml==0.8.1
realtime==1.0.0
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
Scrapy==2.10.0
selenium==4.11.2
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
stack-data==0.6.2
storage3==0.5.4
streamlit==1.26.0
streamlit-folium==0.13.0
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
sympy==1.12
tenacity==8.2.3
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tiktoken==0.4.0
tinysegmenter==0.3
tldextract==3.4.4
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.1
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
Unidecode==1.3.6
urllib3==1.26.16
validators==0.21.2
w3lib==2.1.2
warcio==1.7.4
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.28
zipp==3.16.2
zope.interface==6.0